Application/requirements.txt
django
django-crispy-forms
whitenoise
pillow
